pyproject.toml
[project]
name = "sp-obs"
version = "0.5.0"
description = "Observability integration with Spinal"
readme = "README.md"
authors = [
    { name = "Andrew van Rensburg", email = "andrew@withspinal.com" }
]
license = "MIT"
requires-python = ">=3.11"
keywords = ["opentelemetry", "observability", "tracing", "monitoring", "llm", "ai", "spinal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "opentelemetry-api>=1.38.0,<2",
    "opentelemetry-sdk>=1.38.0,<2",
    "httpx>=0.28.0,<1.0.0",
    "orjson>=3.11.1,<4",
    "opentelemetry-instrumentation-httpx>=0.57b0,<1",
    "opentelemetry-instrumentation-requests>=0.57b0,<1",
    "opentelemetry-semantic-conventions-ai>=0.4.12",
    "aiohttp>=3.13.0",
    "opentelemetry-instrumentation-aiohttp-client>=0.58b0",
]

[project.urls]
Homepage = "https://github.com/withspinal/sp-obs"
Issues = "https://github.com/withspinal/sp-obs/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
exclude = [
    "/tests",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "/tests",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.5",
    "twine>=6.1.0",
]

test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.0",
    "requests-mock>=1.12.1",
    "httpx>=0.28.0",
    "requests>=2.32.0",
]
